feat: Added components for adding, displaying, and viewing posts
Created the AddPost component to allow users to add new posts. Implemented a modal dialog using MUI's Modal component to provide a user-friendly interface for adding posts.

Introduced the Posts component to display a list of posts.

Implemented the Post component to display individual posts.

Added a Route to the Post component in the App.jsx file, enabling navigation to the individual post view.

Incorporated the AddPost component into the Sidenavbar.jsx and BottomNavbar.jsx files for easy access to create posts.